Project cleanup and path fix
Cleaned the forked project of unneccessary temp files, added a .gitignore
and applied the hotfix for sass imports ->
The original version concatenated paths to an import statement with a backslash
in Windows environments. This leads to an error in scss-compile.
Fixed it qnd by replacing any backslashes with forward slashes